Move TransferredCoinInterface
Move TransferredCoinInterface from wallets/index.tsx to src/interface/token/TransferredCoin.ts because its used all around the project.
