+++ src/components/Wallets/Details/SelectedWalletDetails.tsx
@@ -5,7 +5,7 @@ import IconEthereum from "/public/assets/icons/ethereum.svg?jsx";
 import IconWallet from "/public/assets/icons/wallets/wallet.svg?jsx";
 // import IconLoading from "/public/assets/icons/wallets/loading.svg?jsx";
 import { TokenRowWallets } from "~/components/Tokens/TokenRowWallets";
-import { type TransferredCoinInterface } from "~/routes/app/wallets";
+import { type TransferredCoinInterface } from "~/interface/token/TransferredCoin";
 
 interface SelectedWalletProps {
   selectedWallet: Signal<WalletTokensBalances | null>;
