+++ src/interface/token/TransferredCoin.ts
@@ -1,0 +1,4 @@
+export interface TransferredCoinInterface {
+    symbol: string;
+    address: string;
+}
